# 00-PROJECTS/HAIDA/haida-main/dev/docs/API/openapi.yaml
openapi: 3.0.3
info:
  title: HAIDA API
  description: |
    HAIDA - Quality Assurance Platform with AI Integration

    A comprehensive testing and auditing platform that combines traditional QA automation
    with AI-powered insights for enterprise software quality assurance.

    ## Key Features
    - Multi-tenant architecture with RBAC
    - Automated testing execution and monitoring
    - AI-powered chat and assistance (Copilot Studio)
    - Comprehensive documentation management
    - Real-time analytics and reporting
    - Enterprise SSO integration (Microsoft Entra ID)
  version: 2.0.0
  contact:
    name: HAIDA Development Team
    email: hola@stayarta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://haida.vercel.app/api/v2
    description: Production server
  - url: http://localhost:8000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - entraAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    entraAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes:
            User.Read: Read user profile
            Files.Read.All: Read files
            Calendars.Read: Read calendar
            Mail.Read: Read mail
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    ChatThread:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        tenant_id:
          $ref: '#/components/schemas/UUID'
        project_id:
          $ref: '#/components/schemas/UUID'
        user_id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        provider:
          type: string
          enum: [copilot-studio, openai, anthropic]
          default: copilot-studio
        thread_id:
          type: string
        status:
          type: string
          enum: [active, archived, deleted]
          default: active
    ChatMessage:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        thread_id:
          $ref: '#/components/schemas/UUID'
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        content_type:
          type: string
          enum: [text, markdown, json, error]
          default: text
        attachments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
              name:
                type: string

    # Reports
...
  parameters:
    tenantId:
      name: tenant_id
      in: header
      description: Tenant/workspace identifier (UUID or slug)
      required: false
      schema:
        type: string
    page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
paths:
...
  /chat/threads:
    get:
      summary: List chat threads
      tags: [chat]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Chat threads retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatThread'

    post:
      summary: Create chat thread
      tags: [chat]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                project_id:
                  $ref: '#/components/schemas/UUID'
      responses:
        '201':
          description: Chat thread created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatThread'
  /chat/threads/{threadId}/messages:
    get:
      summary: Get thread messages
      tags: [chat]
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'

    post:
      summary: Send message to thread
      tags: [chat]
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                attachments:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [file, image, link]
                      url:
                        type: string
                        format: uri
                      name:
                        type: string
      responses:
        '201':
          description: Message sent and processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_message:
                    $ref: '#/components/schemas/ChatMessage'
                  assistant_response:
                    $ref: '#/components/schemas/ChatMessage'
                  thread_id:
                    $ref: '#/components/schemas/UUID'

  # Reports
...